+++ changelog.md
@@ -13,6 +13,7 @@ If you are reading this on the website, then these changes are live here.
 ### Interpreter
 - Fix some bugs and crashes
 - Improve some formatting with multiline functions and switch functions
+- [`identity` `∘`](https://uiua.org/docs/identity) no longer formats from `()`
 ### Website
 - Allow disabling autorun for pad links